+++ hardware/standards/chessboardstandard.h
@@ -25,9 +25,9 @@ public:
         \param scale The length of a square's side, in millimeters. */
     ChessboardStandard(cv::Size size, double scale);
 
-    std::vector<cv::Point2d> getImagePoints(cv::Mat image, bool *success);
+    std::vector<cv::Point2f> getImagePoints(cv::Mat image, bool *success);
 
-    std::vector<cv::Point3d> getObjectPoints();
+    std::vector<cv::Point3f> getObjectPoints();
 
 
 };
